Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/combine-multiple-policies-expected.txt

CONSOLE MESSAGE: line 11: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 14: Refused to apply a stylesheet because its hash, its nonce, or 'unsafe-inline' does not appear in the style-src directive of the Content Security Policy.
This test checks that we enforce all the supplied policies. This test passes if it doesn't alert fail and if the style doesn't apply.
